irt_bot/requirements.txt
openai>=1.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
slack-bolt>=1.18.0
